Move profile import to own activity
