+++ brouter-routing-app/src/main/AndroidManifest.xml
@@ -16,37 +16,48 @@
         android:label="@string/app_name"
         android:allowBackup="false"
         android:preserveLegacyExternalStorage="true">
-        <activity android:name=".BRouterActivity"
+        <activity
+            android:name=".BRouterActivity"
             android:label="@string/app_name"
             android:exported="true"
-            android:screenOrientation="unspecified" 
+            android:screenOrientation="unspecified"
             android:theme="@android:style/Theme.NoTitleBar.Fullscreen">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".BInstallerActivity"
+            android:label="@string/app_name"
+            android:screenOrientation="landscape"
+            android:launchMode="singleTask"
+            android:exported="true"
+            android:theme="@android:style/Theme.NoTitleBar.Fullscreen"></activity>
+        <activity
+            android:name=".BImportActivity"
+            android:label="BRouter Profile Import"
+            android:exported="true"
+            android:theme="@android:style/Theme">
             <!-- some apps (bluemail) do not recognize the .brf file extention, startactivity+intent is done for attachement with text/plain -->
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
                 <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="text/plain"/>
+                <data android:mimeType="text/plain" />
             </intent-filter>
 
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-                <category android:name="android.intent.category.MULTIWINDOW_LAUNCHER" />
-            </intent-filter>
             <!-- general intent to register .brf files -->
             <intent-filter
                 android:label="@string/app_name"
                 android:priority="50">
                 <action android:name="android.intent.action.VIEW" />
+
                 <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.BROWSABLE" />
-                <data android:scheme="file"/>
-                <data android:scheme="content"/>
-                <data android:host="*"/>
+
+                <data android:scheme="file" />
+                <data android:scheme="content" />
+                <data android:host="*" />
                 <data android:pathPattern=".*\\.brf" />
                 <data android:pathPattern=".*\\..*\\.brf" />
                 <data android:pathPattern=".*\\..*\\..*\\.brf" />
@@ -58,12 +69,14 @@
                 android:label="@string/app_name"
                 android:priority="50">
                 <action android:name="android.intent.action.VIEW" />
+
                 <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.BROWSABLE" />
-                <data android:scheme="file"/>
-                <data android:scheme="content"/>
-                <data android:host="*"/>
-                <data android:mimeType="*/*"/>
+
+                <data android:scheme="file" />
+                <data android:scheme="content" />
+                <data android:host="*" />
+                <data android:mimeType="*/*" />
                 <data android:pathPattern=".*\\.brf" />
                 <data android:pathPattern=".*\\..*\\.brf" />
                 <data android:pathPattern=".*\\..*\\..*\\.brf" />
@@ -71,23 +84,17 @@
                 <data android:pathPattern=".*\\..*\\..*\\..*\\..*\\.brf" />
             </intent-filter>
         </activity>
-        <activity android:name=".BInstallerActivity"
-            android:label="@string/app_name"
-            android:screenOrientation="landscape"
-            android:launchMode="singleTask"
-            android:exported="true"
-            android:theme="@android:style/Theme.NoTitleBar.Fullscreen">
-        </activity>
+
         <service
             android:exported="true"
             android:name=".BRouterService"
             android:enabled="true"
             android:process=":brouter_service" />
-        <service android:name="btools.routingapp.DownloadService"
+        <service
+            android:name="btools.routingapp.DownloadService"
             android:label="Download Service"
             android:icon="@mipmap/ic_launcher"
-            android:enabled="true"
-            />
+            android:enabled="true" />
 
     </application>
 </manifest>
